wrap forward shading pipeline
