+++ opengl_rt.vcxproj
@@ -132,6 +132,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="src\AO.cpp" />
+    <ClCompile Include="src\Application.cpp" />
     <ClCompile Include="src\global.cpp" />
     <ClCompile Include="src\ImGUIManager.cpp" />
     <ClCompile Include="src\imgui_impl_glfw.cpp" />
@@ -142,6 +143,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\AO.h" />
+    <ClInclude Include="src\Application.h" />
     <ClInclude Include="src\Camera.h" />
     <ClInclude Include="src\global.h" />
     <ClInclude Include="src\ImGUIManager.h" />
