+++ opengl_rt.vcxproj.filters
@@ -39,6 +39,9 @@
     <ClCompile Include="src\global.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
+    <ClCompile Include="src\Application.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\Shader.h">
@@ -87,6 +90,9 @@
       <Filter>头文件</Filter>
     </ClInclude>
     <ClInclude Include="src\global.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
+    <ClInclude Include="src\Application.h">
       <Filter>头文件</Filter>
     </ClInclude>
   </ItemGroup>
